fetch: изменение типа id на Long

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -16,7 +16,7 @@ public class FilmController {
 
     private final List<Film> films = new ArrayList<>();
     private final FilmValidator validator = new FilmValidator();
-    private static int startId = 1;
+    private static Long startId = 1L;
 
     @GetMapping
     public List<Film> getFilms() {
